finished hike/create action

+++ db/migrate/20170621071131_devise_create_users.rb
@@ -2,7 +2,7 @@ class DeviseCreateUsers < ActiveRecord::Migration[5.0]
   def change
     create_table :users do |t|
       t.integer :total_elevation_climbed, null: false, default: "0"
-      t.integer :total_distance_hiked, null: false, default: "0"
+      t.float :total_distance_hiked, null: false, default: "0"
       t.integer :total_hikes, null: false, default: "0"
 
       ## Database authenticatable
